bring test articles up to date

+++ src/timesync/fuzz-nullcipher.c
@@ -12,9 +12,9 @@
 #define BLKSIZ 16
 
 int NTS_encrypt(uint8_t *ctxt,
-                int ctxt_len,
+                size_t ctxt_len,
                 const uint8_t *ptxt,
-                int ptxt_len,
+                size_t ptxt_len,
                 const AssociatedData *info,
                 const NTS_AEADParam *aead,
                 const uint8_t *key) {
@@ -32,9 +32,9 @@ int NTS_encrypt(uint8_t *ctxt,
 }
 
 int NTS_decrypt(uint8_t *ptxt,
-                int ptxt_len,
+                size_t ptxt_len,
                 const uint8_t *ctxt,
-                int ctxt_len,
+                size_t ctxt_len,
                 const AssociatedData *info,
                 const NTS_AEADParam *aead,
                 const uint8_t *key) {
